Split BoardState to other files
